requirements.txt
dash>=2.15.0
dash-bootstrap-components>=1.5.0
plotly>=5.17.0
numpy>=1.24.0
opencv-python>=4.8.0
pillow>=10.0.0